+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## TBC
+
+### Fixed
+
+- (react-native) Fix reporting of `RCTFatal()` crashes on iOS. [#1719](https://github.com/bugsnag/bugsnag-js/pull/1719)
+
 ## v7.16.3 (2022-04-05)
 
 ### Changed